refactor: 네트워크 연결 감지 및 UI 로직을 계층별로 분리 (#182)
